requirements.txt
dacite>=1.6.0
descartes>=1.1.0
matplotlib>=3.4.3
numpy>=1.20.3
pandas>=1.4.4
Pillow>=9.2.0
plotly>=5.9.0
scipy>=1.7.3
setuptools>=58.0.4
Shapely>=1.8.1.post1
typing-extensions>=4.1.1
